major: Fix critical Frotz compatibility issue - player object creation
BREAKTHROUGH: Resolved root cause from Frotz debug session analysis

Key Fixes:
- Create player object as Z-Machine object #1 automatically
- Fix property number collision (desc=1, location=2)
- Correct Z-Machine v3 property format encoding
- Add essential player properties to property registry
- Implement proper dual numbering system (IR ID vs Z-Machine object numbers)

Architecture Documentation:
- Add comprehensive object numbering system documentation
- Document property format specification compliance
- Update Grue compiler architecture guide

Progress:
- Before: "Fatal error: @get_prop called with object 0"
- After: Basic examples work perfectly with Frotz
- Remaining: Secondary property table reading issue (both interpreters)

This resolves the fundamental missing-player-object issue that was
blocking external Z-Machine interpreter compatibility.

+++ src/instruction.rs
@@ -140,13 +140,13 @@ impl Instruction {
         let form = match opcode_byte >> 6 {
             0b11 => InstructionForm::Variable,
             0b10 => InstructionForm::Short,
-            0b01 => InstructionForm::Long, // REVERT: 01 is Long form, not Short
+            0b01 => InstructionForm::Long,
             0b00 => {
                 // Check for extended form (0xBE in V5+)
                 if opcode_byte == 0xBE && version >= 5 {
                     InstructionForm::Extended
                 } else {
-                    InstructionForm::Long
+                    InstructionForm::Short // FIX: 00 is Short form, not Long
                 }
             }
             _ => unreachable!(),
